Evict lights with preserveCachedShadow if they are disabled while pending placement
Jira: https://jira.unity3d.com/browse/UUM-90709

This PR addresses a reverses the changes made in https://github.cds.internal.unity3d.com/unity/unity/pull/56366 and provides an alternative way to fix the crash.

When an HDAdditionalLightData is disabled, it contains an HDLightRenderEntity which is destroyed. When preserveCachedShadow is set to true, since the light is not evicted from the Atlas, In HDCachedShadowAtlas.m_RegisteredLightDataPendingPlacement, we end up with an entry that points to the older, now destroyed HDLightRenderEntity, causing a crash.

This PR fixes this by ensuring that even when preserveCachedShadow is set, we evict the light if it has not been yet placed in the atlas.

*[The PR comment commands commands and workflows to ready and land PR's can be found in the [pipeline documentation](https://internaldocs.unity.com/editor_and_runtime_development_guide/DevelopmentProcess/pr-process/evergreen-merge-queue/#rejoin-the-queue).]*

*[Description of feature/change. Links to screenshots, design docs, user docs, etc. Link to Jira if applicable. Remember reviewers may be outside your team, and not know your feature/area that should be explained more.]*