Fix broken GI HDRP_Tests
This PR is the first in an initiative to improve the health of our graphics tests.

HDRP_Tests had a label on certain scenes. If this label is present, the test runner baked GI while running the test. There was some custom code in the test project that was switching the baking backend. This code was failing, causing the HDRP_Tests job to fail on MacOS. The GI team has made a change, where baking is not supposed to be done at test-time for these tests; they are only meant to test runtime lightmap sampling behavior (as opposed to baking behavior).

I've removed the offending script after consulting the GI team. I've deleted the label from the scene names, so GI is not live-baked any more. I've checked-in the baked light data. In some places, this has involved changing the scene-specific gitignore file.

After landing this PR, the HDRP_Tests projects loads successfully on MacOS, and tests start failing individually (which is better than errors while loading the project). Next steps will to start resolving this.