[SRP] Use platform defaults instead of hard-coded 32bits depth [PLATGRAPH-4062]
There are many places in SRP (URP, HDRP, Rendergraph) that default to a hard-coded `D32_SFloat` or `D32_SFloat_S8_UInt` format when it comes to deciding on a Depth format.

For some platforms this has some negative consequences on performance.

To improve performance for such platforms, this pull request replaces hard-coded usage of D32_SFloat in SRP code, and replaces it with a platform-specific value.
(The function that returns that value is in CoreUtils)





Jira ticket: [PLATGRAPH-4062](https://jira.unity3d.com/browse/PLATGRAPH-4062)