[Unity 6.1][URP] Fixing aliased shadows due to lack of UV precision (UUM-87935)
Fixes UUM-87935.
The problem is that the sampling functions are using half, which doesn't have enough precision for the 2K, 4K and 8K shadowmaps that URP offers. The solution is to create a version of those functions using float instead of half.

New functions were added instead of changing the current ones to not break users who might be relying on those.