[VFX][Fix] Allow zero-initializing strip data buffer with a compute dispatch
There was a missing variant for the zero initialization kernel that prevented initializing the strip data buffer using a dispatch. This could be problematic for systems with a high strip count because it meant that we were inefficiently passing and copying a large cpu buffer through the device layers up to the GPU.

This PR adds this variant in the `VFXCopyBuffer.compute` compute shader file, and uses it when the size of the buffer to initialize is larger than 64.