Graphics/SRP/RPF - Render Graph - Setting ResourceHandle as readonly
This is a small internal optimization in Render Graph codebase.

While reviewing #57093, I realized that we were generating defensive copy of `ResourceHandle` in many places by calling its `index` property getter. C# runtime generates a copy when the `ResourceHandle` instance is passed with `in` keyword because it doesn't know if the `index` getter is modifying the instance or not.

To counter this, `ResourceHandle` is now a `readonly struct`, we cannot modify its version anymore as we used to do but the version is bumped only twice in the code in a specific function. So we can just assign a new `ResourceHandle` instance with the right version in those two places. It is a good trade-off as we should avoid many defensive copies in the rest of the RenderGraph codebase. Also `ResourceHandle` is internal so no risk of a breaking change.