UUM-91854 - Fix lightmap index changes when loading / unloading scenes with GPU Resident Drawer
When scenes containing lightmapped renderers are loaded additively or unloaded, the lightmapping manager reshuffles the lightmap indices to accommodate the new lightmaps loaded in the scene's lighting data. Unfortunately, when using the GPU Resident Drawer, these new lightmap indices where not taken into account in the GRD's instance data, resulting in objects potentially using the incorrect lightmap after unloading of other scenes. This is because the GRD relies on the ObjectDispatcher to be notified in any changes in scene data. But the ObjectDispatcher tracks changes only if objects are marked as "dirty" after modification, which wasn't the case when the lightmap indices where modified after load / unload of scenes.

The PR fixes that by adding an extra method in Renderer for setting lightmap indices with an extra `SetDirty` call so these modifications get tracked by the ObjectDispatcher. The dirty flag is set ONLY if the lightmap indices have really changed, though. Similarly, for symmetry, `SetDirty` was also added in `Terrain::SetLightmapIndices`.

This is potentially not optimal for the GRD because the ObjectDispatcher doesn't distinguish between newly created objects and modified objects so any changes of lightmaps indices gets "seen" by the GRD as a newly created mesh renderer so results in a tear down and reconstruction of instance data for that renderer. In the future, especially in the case of streaming where we would want to minimize CPU cost of adding / removing objects, it might be interesting to either modify ObjectDispatcher so we can distinguish between creation & changes in objects, and even be finer grained in what has changed in terms of properties.

The PR also adds a new GRDLightMap graphics test that verifies the issue with additive scene loading / unloading in the case of GRD with URP.