Render Graph Pass Merging Leads to Unnecessary Clear Load Actions
When merging Render Graph Passes into a single Native Render Pass, the accessFlags of the resources are also merged by an OR operation to define the Load and Store operations.
Currently if we merge a two passes accessing the same resource, first as `WriteAll` and second as `Read`, are generating an internal merged access flag as `Read|Write|Discard`, the same result as it was at the opposite order, `Read`+`WriteAll`. But, in the first case, the Load operation should be `DontCare` as the pass will write all pixels and don't need the info coming from other passes, and in the second case, it will need to make a `Load` operation.

With the change in this PR, when merging the passes and OR'ing the accessFlags, it will check if it already has a `Discard` (`WriteAll` is `Write +  Discard`) for not adding new flags as the content should be discarded in anyway.

![image](https://media.github.cds.internal.unity3d.com/user/7938/files/7bfaeb41-1700-4f72-8bf8-7b7c66f26a39)