[Port] [6000.0] [Rendering Debbugger] Avoid  reflection code and use directly the volume parameters of the Volume Component.
The Volume Debugger page was using reflection to obtain the volume parameters from the selected volume component in the stack. This code as was copy pasted from VolumeComponent and Volume Component editor was not taking into account the logic of .

- Getting public fields without NonSerialize attribute
- Getting private fields with SerializeField.

The selected component of the jira, had an RTHandle that was doing an infinte loop, that was making a crash directly in mono.

I've gotten rid of all this logic that was copy pasted, and instead I am using the Volume Parameter list, but I've added debug info to it.

Making that, it allowed me to simplify the table generation. Having a single code that

1- Gets the chain of volumes
2- Generates the rows of the table
3- For each volume in the interpolation chain, we create a column.

The code now is generic, instead of having granular cases for defaults, customs, etc.

I've improved and simplified the code, and I've added alternate colors to the table.

![image](https://github.cds.internal.unity3d.com/unity/unity/assets/3279/67b2aceb-dd18-40a0-8941-af7f7d4c7b28)

Also, the values that are not interpolated show now "-", making it easier to see where the values are overriden.