[Port] [6000.0] SpeedTree does not move when using WindZone
This PR fixes the _WindEnabled implementation as a local variable.

In the older engine versions, the speedtree shader received _WindEnabled variable, the variable hasn't passed to the shader since Unity6. Instead, it is fixed to create the windEnabled variable as a local variable in this PR.

This PR covers both SpeedTree 7 and 8.