[Rendering Debugger]Moving resources out from resource folder and put them on Graphics Settings.
The Rendering Debugger Prefabs are is using uGUI, and if a project is using UITK, the code and dependencies from uGUI. get pulled into our builds because of that reference.

Also in retail build size, the References for the Rendering Debugger are included. With this change, those are not.

UUM-90292