[Port] [6000.0] [HDRP] [Wizzard] Fix default state 'Fail' used while waiting from asynchronous check
Fix [UUM-85253](https://jira.unity3d.com/browse/UUM-85253). Due to the delayed response from package management, the status displayed between the fix and the checking of the fix was the default "fail". So it was really strange. And this was enforced even by the unexpectedly long delay for PackageManager to answer.

To solve this, this PR introduce a "pending" state that is used while we are waiting for the answer instead of assuming it is OK or KO.

![Unity_cYtolKRenN](https://github.cds.internal.unity3d.com/unity/unity/assets/465/763ae0df-6708-4ba2-8888-c3bedb06cd0f)
(wait it is a little long to get answer from PackageManager)