[URP] Fixing an issue with fading baked shadows when Real-Time Additional Shadows are disabled (UUM-93148)
Fixes an issue when Real-Time Additional Light Shadows are disabled but using baked shadows.
The issue is that `_AdditionalShadowParams` is not included nor updated when real-time shadows are disabled.
This PR reverts the behaviour to what it was before. A follow-up PR may come to add the feature for baked shadows only.