[UUM-90909] Fix artifacts when blending cascade shadows and distance shadowmask
This [PR](https://github.cds.internal.unity3d.com/unity/unity/pull/53502) uses the camera-based blending weight used in URP. Blending weight means the ratio of shadowmask. HDRP has a border value for each cascade, and the blending weight is calculated based on the cascade sphere position.

The blending weight is visualized as follows.

* HDRP
![image](https://github.cds.internal.unity3d.com/unity/unity/assets/1641/c79d0a0e-2172-4422-9d83-aa4610103d01)

* URP
![image](https://github.cds.internal.unity3d.com/unity/unity/assets/1641/574c8a4b-60d3-4353-8ba0-c1f7e561f0d2)

The blending weight also affects the sampled shadow value.

The artifact issue occurred when calculating shadow attenuation by using the blending weight based on the camera position and the shadow value based on the cascade sphere position together.

The blending weight was reverted to be calculated based on the cascade sphere. Instead, the blending weight based on the camera position was used to replace the area where the shadow value could not be obtained with a shadow mask.

The problem of not obtaining the shadow value is fundamentally a problem with culling. It occurs when the camera position is fixed and the camera looks in a specific direction near the max shadow distance.