Fix for distorted SplashScreen scene when using eye texture dynamic resolution
Fix for [UUM-93163](https://jira.unity3d.com/browse/UUM-93163).
When using eye texture dynamic resolution it was reported that the splash screen looked distorted if the scale is smaller than 1. The reason for it is because the camera scene doesn't have the URP Dynamic Resolution flag enabled, but the eye texture dynamic resolution uses XRSystem flag to be enabled and the `cameraData.allowdsDynamicResolution` prevent the eye texture to be scaled properly.

The remove of the flag fixes the issue